refactored gannt chart, refactored code into temporary helper modules, protected all project views base on views

+++ app/helpers/invitations_helper.rb
@@ -8,8 +8,9 @@ module InvitationsHelper
     render :partial => 'invitations/fields', :locals => { :f => f }
   end
   
-  def list_invitations(invitations)
-    render :partial => 'invitations/invitation', :collection => invitations
+  def list_invitations(project,invitations)
+    render :partial => 'invitations/invitation', :collection => invitations,
+    :locals => { :project => project }
   end
 
   def list_pending_invites(invitations)
@@ -17,6 +18,7 @@ module InvitationsHelper
   end
   
   def delete_invitation_link(invitation)
+    return unless invitation.editable?(current_user)
     link_to_remote trash_image,
       :url => project_invitation_path(invitation.project,invitation),
       :method => :delete
@@ -26,10 +28,11 @@ module InvitationsHelper
     link_to 'Invite someone', new_project_invitation_path(project)
   end
   
-  def resend_invitation_link(invitation)
+  def resend_invitation_link(project,invitation)
+    return unless invitation.editable?(current_user)
     link_to_remote t('.resend'),
-      :url => resend_project_invitation_path(invitation.project,invitation),
-      :loading => show_loading('resend_invitation', invitation.id),
+      :url => resend_project_invitation_path(project,invitation),
+      :loading => show_loading('resend_invitation',invitation.id),
       :html => { :id => "resend_invitation_#{invitation.id}_link" }
   end
   
@@ -37,5 +40,13 @@ module InvitationsHelper
     page.replace "resend_invitation_loading_#{invitation.id}",
       :partial => 'invitations/sent'
   end
+
+  def invite_form(project,invitation)
+    return unless project.editable?(current_user)
+    render :partial => 'invitations/new', :locals => {
+      :project => project,
+      :invitation => invitation }
+  end
+
   
 end
